MAINT: stats: move permutation_test to _resampling.py
